+++ src/config/ConfigServer.cpp
@@ -7,7 +7,7 @@ ConfigServer::~ConfigServer()
 {}
 
 /* -- Setter -- */
-void	ConfigServer::setPort(int16_t port)						{ _port = port; }
+void	ConfigServer::setPort(int port)							{ _port = port; }
 void	ConfigServer::setHost(const std::string& host)			{ _host = host; }
 void	ConfigServer::addServerName(const std::string& name)	{ _server_names.push_back(name); }
 
@@ -23,7 +23,7 @@ void	ConfigServer::addLocation(const ConfigLocation& loc)
 
 
 /* -- Getters -- */
-int16_t								ConfigServer::getPort() const			{ return (_port); }
+int									ConfigServer::getPort() const			{ return (_port); }
 std::string							ConfigServer::getHost() const			{ return (_host); }
 const std::vector<std::string>&		ConfigServer::getServerNames() const	{ return (_server_names); }
 const std::vector<ConfigLocation>&	ConfigServer::getLocations() const		{ return (_locations); }
